+++ src/main/java/chess/controller/ChessController.java
@@ -1,40 +1,76 @@
 package chess.controller;
 
-import chess.domain.ChessGame;
-import chess.domain.command.Command;
 import chess.domain.command.Commands;
-import chess.view.InputView;
-import chess.view.OutputView;
+import chess.domain.exception.DataException;
+import chess.domain.vo.MoveVo;
+import chess.service.ChessService;
+import chess.view.ModelView;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
 
+import java.sql.SQLException;
+
+@Controller
+@RequestMapping("/play")
 public class ChessController {
 
-    public void run() {
-        ChessGame chessGame = new ChessGame();
-        OutputView.printCommandInfo();
-        while (chessGame.isRunning()) {
-            final Commands inputCmd = InputView.inputCommand();
-            Command command = Command.of(inputCmd.mainCommand());
-            command.apply(chessGame, inputCmd);
-            view(chessGame, command);
+    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
+    private final ChessService chessService;
+
+    public ChessController(ChessService chessService) {
+        this.chessService = chessService;
+    }
+
+    @GetMapping("")
+    public String play(Model model) throws DataException {
+        model.addAllAttributes(ModelView.startResponse(chessService.loadHistory()));
+        return "play";
+    }
+
+    @GetMapping("/new")
+    public String playNewGameWithNoSave(Model model) {
+        model.addAllAttributes(ModelView.newGameResponse(chessService.initialGameInfo()));
+        return "chessGame";
+    }
+
+    @GetMapping("/{name}/new")
+    public String playNewGameWithSave(Model model, @PathVariable String name) throws DataException {
+        model.addAllAttributes(ModelView.newGameResponse(
+                chessService.initialGameInfo(),
+                chessService.addHistory(name)
+        ));
+        return "chessGame";
+    }
+
+    @GetMapping("/continue")
+    public String continueGame(Model model, @RequestParam("name") String name) throws DataException {
+        final String id = chessService.getIdByName(name);
+        model.addAllAttributes(ModelView.commonResponseForm(chessService.continuedGameInfo(id), id));
+        return "chessGame";
+    }
+
+    @GetMapping("/end")
+    public String endGame() {
+        return "play";
+    }
+
+    @PostMapping(path = "/move")
+    @ResponseBody
+    public String move(@RequestBody MoveVo moveVo) {
+        String command = makeMoveCmd(moveVo.getSource(), moveVo.getTarget());
+        String historyId = moveVo.getGameId();
+        try {
+            chessService.move(historyId, command, new Commands(command));
+            return GSON.toJson(ModelView.moveResponse(chessService.continuedGameInfo(historyId), historyId));
+        } catch (IllegalArgumentException | SQLException e) {
+            return e.getMessage();
         }
     }
 
-    private void view(ChessGame chessGame, Command command) {
-        if (command.isStart()) {
-            OutputView.printBoard(chessGame.boardDto());
-        }
-        if (command.isMove()) {
-            OutputView.printBoard(chessGame.boardDto());
-            confirmKingDead(chessGame);
-        }
-        if (command.isStatus()) {
-            OutputView.printStatus(chessGame.pointDto());
-        }
-    }
-
-    private void confirmKingDead(ChessGame chessGame) {
-        if (chessGame.isEnd()) {
-            OutputView.printWinner(chessGame.winner());
-        }
+    private String makeMoveCmd(String source, String target) {
+        return String.join(" ", "move", source, target);
     }
 }
